feat(ingest): loosen sqlalchemy dep & support airflow 2.3+ (#6204)
