+++ metadata-ingestion-modules/airflow-plugin/setup.py
@@ -1,8 +1,8 @@
 import os
 import pathlib
-from typing import Dict, Set
 
 import setuptools
+
 
 package_metadata: dict = {}
 with open("./src/datahub_airflow_plugin/__init__.py") as fp:
@@ -23,9 +23,7 @@ base_requirements = {
     "typing-inspect",
     "pydantic>=1.5.1",
     "apache-airflow >= 2.0.2",
-    "acryl-datahub[airflow] >= 0.8.36",
-    # Pinned dependencies to make dependency resolution faster.
-    "sqlalchemy==1.3.24",
+    f"acryl-datahub[airflow] == {package_metadata['__version__']}",
 }
