+++ metadata-ingestion/docs/sources/oracle/oracle.md
@@ -1,1 +1,1 @@
-As a SQL-based service, the Athena integration is also supported by our SQL profiler. See here for more details on configuration.
+As a SQL-based service, the Oracle integration is also supported by our SQL profiler. See here for more details on configuration.
