+++ metadata-ingestion/setup.cfg
@@ -22,7 +22,7 @@ ban-relative-imports = true
 
 [mypy]
 plugins =
-    sqlmypy,
+    ./tests/test_helpers/sqlalchemy_mypy_plugin.py,
     pydantic.mypy
 exclude = ^(venv|build|dist)/
 ignore_missing_imports = yes
@@ -55,6 +55,7 @@ disallow_untyped_defs = yes
 asyncio_mode = auto
 addopts = --cov=src --cov-report term-missing --cov-config setup.cfg --strict-markers
 markers =
+    airflow: marks tests related to airflow (deselect with '-m not airflow')
     slow_unit: marks tests to only run slow unit tests (deselect with '-m not slow_unit')
     integration: marks tests to only run in integration (deselect with '-m "not integration"')
     integration_batch_1: mark tests to only run in batch 1 of integration tests. This is done mainly for parallelisation (deselect with '-m not integration_batch_1')
