+++ metadata-ingestion/src/datahub/ingestion/source/sql/trino.py
@@ -4,15 +4,12 @@ from textwrap import dedent
 from typing import Any, Dict, List, Optional
 
 import sqlalchemy
-
-# This import verifies that the dependencies are available.
-import trino.sqlalchemy  # noqa: F401
 from pydantic.fields import Field
 from sqlalchemy import exc, sql
 from sqlalchemy.engine import reflection
 from sqlalchemy.engine.reflection import Inspector
 from sqlalchemy.sql import sqltypes
-from sqlalchemy.sql.type_api import TypeEngine
+from sqlalchemy.types import TypeEngine
 from trino.exceptions import TrinoQueryError
 from trino.sqlalchemy import datatype, error
 from trino.sqlalchemy.dialect import TrinoDialect
