+++ metadata-ingestion/src/datahub/ingestion/source/usage/redshift_usage.py
@@ -3,13 +3,12 @@ import dataclasses
 import logging
 import time
 from datetime import datetime
-from typing import Dict, Iterable, List, Optional, Set
+from typing import TYPE_CHECKING, Dict, Iterable, List, Optional, Set
 
 from pydantic.fields import Field
 from pydantic.main import BaseModel
 from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
-from sqlalchemy.engine.result import ResultProxy, RowProxy
 
 import datahub.emitter.mce_builder as builder
 from datahub.configuration.source_common import EnvBasedSourceConfigBase
@@ -38,6 +37,13 @@ from datahub.metadata.schema_classes import (
 )
 
 logger = logging.getLogger(__name__)
+
+if TYPE_CHECKING:
+    try:
+        from sqlalchemy.engine import Row  # type: ignore
+    except ImportError:
+        # See https://github.com/python/mypy/issues/1153.
+        from sqlalchemy.engine.result import RowProxy as Row  # type: ignore
 
 REDSHIFT_DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S"
 
@@ -266,7 +272,7 @@ class RedshiftUsageSource(Source):
         logger.debug(f"sql_alchemy_url = {url}")
         return create_engine(url, **self.config.options)
 
-    def _should_process_row(self, row: RowProxy) -> bool:
+    def _should_process_row(self, row: "Row") -> bool:
         # Check for mandatory proerties being present first.
         missing_props: List[str] = [
             prop
@@ -294,10 +300,13 @@ class RedshiftUsageSource(Source):
     def _gen_access_events_from_history_query(
         self, query: str, engine: Engine
     ) -> Iterable[RedshiftAccessEvent]:
-        results: ResultProxy = engine.execute(query)
-        for row in results:  # type: RowProxy
+        results = engine.execute(query)
+        for row in results:
             if not self._should_process_row(row):
                 continue
+            if hasattr(row, "_asdict"):
+                # Compatibility with sqlalchemy 1.4.x.
+                row = row._asdict()
             access_event = RedshiftAccessEvent(**dict(row.items()))
             # Replace database name with the alias name if one is provided in the config.
             if self.config.database_alias:
